requirements.txt
Django==2.1.3
djangorestframework==3.8.2
paho-mqtt==1.4.0
requests==2.20.1
psycopg2-binary==2.7.6.1django-celery-beat==1.3.0
django-crontab==0.7.1
django-crontab==0.7.1
